Hook up Users::GraphQL::Run action
