+++ backend/apps/web/controllers/graphql/run.rb
@@ -1,5 +1,3 @@
-require "controllers/cors"
-
 module Web::Controllers::Graphql
   class Run
     include Web::Action
@@ -51,17 +49,6 @@ module Web::Controllers::Graphql
 
     def verify_csrf_token?
       false
-    end
-
-    def find_current_user(token)
-      return unless token
-
-      token = token.split.last
-      return unless token
-
-      payload, _headers = JWT.decode token, ENV.fetch('AUTH_TOKEN_SECRET'), true, algorithm: 'HS256'
-      user_repo = UserRepository.new
-      user_repo.find_by_email(payload["email"])
     end
   end
 end
