+++ backend/lib/books.rb
@@ -1,2 +1,5 @@
+require_relative "controllers/cors"
+require_relative "controllers/current_user"
+
 module Books
 end
